feat:将运行时agent循环拆分为独立运行时组件
